+++ packages/uniforms-antd/src/ErrorsField.js
@@ -3,7 +3,7 @@ import React          from 'react';
 import filterDOMProps from 'uniforms/filterDOMProps';
 import nothing        from 'uniforms/nothing';
 
-const ErrorsField = ({className, children, ...props}, {uniforms: {error, schema}}) =>
+const ErrorsField = ({children, ...props}, {uniforms: {error, schema}}) =>
     (!error && !children) ? nothing : (
         <div {...filterDOMProps(props)}>
             {children}