+++ packages/uniforms-antd/src/SubmitField.js
@@ -2,7 +2,7 @@ import BaseField from 'uniforms/BaseField';
 import Button    from 'antd/lib/button';
 import React     from 'react';
 
-const SubmitField = ({className, inputRef, value, ...props}, {uniforms: {error, state: {disabled}}}) =>
+const SubmitField = ({inputRef, value, ...props}, {uniforms: {error, state: {disabled}}}) =>
     <Button
         disabled={!!(error || disabled)}
         htmlType="submit"