+++ packages/uniforms-bootstrap4/__tests__/_createSchema.js
@@ -1,0 +1,7 @@
+import SimpleSchema from 'simpl-schema';
+
+import createSchemaBridge from 'uniforms/createSchemaBridge';
+
+const createSchema = schema => createSchemaBridge(new SimpleSchema(schema));
+
+export default createSchema;