+++ packages/uniforms-bootstrap4/src/wrapField.js
@@ -1,9 +1,11 @@
-import React      from 'react';
-import classnames from 'classnames';
+import React          from 'react';
+import classnames     from 'classnames';
+import filterDOMProps from 'uniforms/filterDOMProps';
 
 import gridClassName from './gridClassName';
 
 export default function wrapField ({
+    checkboxes,      // eslint-disable-line no-unused-vars
     className,
     disabled,
     error,
@@ -12,10 +14,15 @@ export default function wrapField ({
     help,            // Help text.
     helpClassName,   // Help text class name.
     id,
+    inline,          // eslint-disable-line no-unused-vars
+    inputClassName,  // eslint-disable-line no-unused-vars
+    inputRef,        // eslint-disable-line no-unused-vars
     label,
     required,
+    rows,            // eslint-disable-line no-unused-vars
     showInlineError, // Show inline error message?
-    wrapClassName    // Input wrapper class name.
+    wrapClassName,   // Input wrapper class name.
+    ...props
 }, children) {
     const hasWrap = !!(grid || wrapClassName);
 
@@ -43,6 +50,7 @@ export default function wrapField ({
                     row: grid
                 }
             )}
+            {...filterDOMProps(props)}
         >
             {label && (
                 <label