+++ packages/uniforms-unstyled/__tests__/RadioField.js
@@ -127,7 +127,6 @@ test('<RadioField> - renders a wrapper with unknown props', () => {
     const element = <RadioField name="x" data-x="x" data-y="y" data-z="z" />;
     const wrapper = mount(element, createContext({x: {type: String, allowedValues: ['a', 'b']}}));
 
-    expect(wrapper.find('div')).toHaveLength(3);
     expect(wrapper.find('div').at(0).prop('data-x')).toBe('x');
     expect(wrapper.find('div').at(0).prop('data-y')).toBe('y');
     expect(wrapper.find('div').at(0).prop('data-z')).toBe('z');